+++ packages/contracts/contracts/voting/addresslist/AddresslistVoting.sol
@@ -6,7 +6,7 @@ import {SafeCastUpgradeable} from "@openzeppelin/contracts-upgradeable/utils/mat
 
 import {IMembership} from "../../core/plugin/IMembership.sol";
 import {IDAO} from "../../core/IDAO.sol";
-import {RATIO_BASE, applyRatioCeiled} from "../../utils/Ratio.sol";
+import {RATIO_BASE, _applyRatioCeiled} from "../../utils/Ratio.sol";
 import {IMajorityVoting} from "../majority/IMajorityVoting.sol";
 import {MajorityVotingBase} from "../majority/MajorityVotingBase.sol";
 import {Addresslist} from "./Addresslist.sol";
@@ -82,6 +82,7 @@ contract AddresslistVoting is IMembership, Addresslist, MajorityVotingBase {
     function createProposal(
         bytes calldata _metadata,
         IDAO.Action[] calldata _actions,
+        uint256 _allowFailureMap,
         uint64 _startDate,
         uint64 _endDate,
         VoteOption _voteOption,
@@ -101,7 +102,8 @@ contract AddresslistVoting is IMembership, Addresslist, MajorityVotingBase {
             _metadata: _metadata,
             _startDate: _startDate,
             _endDate: _endDate,
-            _actions: _actions
+            _actions: _actions,
+            _allowFailureMap: _allowFailureMap
         });
 
         // Store proposal related information
@@ -114,13 +116,18 @@ contract AddresslistVoting is IMembership, Addresslist, MajorityVotingBase {
         proposal_.parameters.snapshotBlock = snapshotBlock;
         proposal_.parameters.votingMode = votingMode();
         proposal_.parameters.supportThreshold = supportThreshold();
-        proposal_.parameters.minVotingPower = applyRatioCeiled(
+        proposal_.parameters.minVotingPower = _applyRatioCeiled(
             addresslistLengthAtBlock(snapshotBlock),
             minParticipation()
         );
 
         // REMOVE
         proposal_.tally.totalVotingPower = addresslistLengthAtBlock(snapshotBlock); // TODO THIS DOESN'T NEED TO BE STORED ANYMORE https://aragonassociation.atlassian.net/browse/APP-1417
+
+        // Reduce costs
+        if(_allowFailureMap != 0) {
+            proposal_.allowFailureMap = _allowFailureMap;
+        }
 
         for (uint256 i; i < _actions.length; ) {
             proposal_.actions.push(_actions[i]);