+++ packages/contracts/contracts/voting/admin/Admin.sol
@@ -53,9 +53,11 @@ contract Admin is IMembershipContract, PluginCloneable, Proposal {
     /// @notice Creates and executes a new proposal.
     /// @param _metadata The metadata of the proposal.
     /// @param _actions The actions to be executed.
+    /// @param _allowFailureMap A bitmap allowing the proposal to succeed, even if individual actions might revert. If the bit at index `i` is 1, the proposal succeeds even if the `i`th action reverts. A failure map value of 0 requires every action to not revert.
     function executeProposal(
         bytes calldata _metadata,
-        IDAO.Action[] calldata _actions
+        IDAO.Action[] calldata _actions,
+        uint256 _allowFailureMap
     ) external auth(EXECUTE_PROPOSAL_PERMISSION_ID) {
         uint64 currentTimestamp64 = block.timestamp.toUint64();
 
@@ -64,8 +66,9 @@ contract Admin is IMembershipContract, PluginCloneable, Proposal {
             _metadata: _metadata,
             _startDate: currentTimestamp64,
             _endDate: currentTimestamp64,
-            _actions: _actions
+            _actions: _actions,
+            _allowFailureMap: _allowFailureMap
         });
-        _executeProposal(dao, proposalId, _actions);
+        _executeProposal(dao, proposalId, _actions, _allowFailureMap);
     }
 }