+++ packages/contracts/contracts/voting/majority/MajorityVotingBase.sol
@@ -132,12 +132,14 @@ abstract contract MajorityVotingBase is
     /// @param tally The vote tally of the proposal.
     /// @param voters The votes casted by the voters.
     /// @param actions The actions to be executed when the proposal passes.
+    /// @param allowFailureMap A bitmap allowing the proposal to succeed, even if individual actions might revert. If the bit at index `i` is 1, the proposal succeeds even if the `i`th action reverts. A failure map value of 0 requires every action to not revert.
     struct Proposal {
         bool executed;
         ProposalParameters parameters;
         Tally tally;
         mapping(address => IMajorityVoting.VoteOption) voters;
         IDAO.Action[] actions;
+        uint256 allowFailureMap;
     }
 
     /// @notice A container for the proposal parameters at the time of proposal creation.
@@ -227,10 +229,10 @@ abstract contract MajorityVotingBase is
     /// @dev This method is required to support [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822).
     /// @param _dao The IDAO interface of the associated DAO.
     /// @param _votingSettings The voting settings.
-    function __MajorityVotingBase_init(
-        IDAO _dao,
-        VotingSettings calldata _votingSettings
-    ) internal onlyInitializing {
+    function __MajorityVotingBase_init(IDAO _dao, VotingSettings calldata _votingSettings)
+        internal
+        onlyInitializing
+    {
         __PluginUUPSUpgradeable_init(_dao);
         _updateVotingSettings(_votingSettings);
     }
@@ -238,9 +240,7 @@ abstract contract MajorityVotingBase is
     /// @notice Checks if this or the parent contract supports an interface by its ID.
     /// @param interfaceId The ID of the interface.
     /// @return bool Returns `true` if the interface is supported.
-    function supportsInterface(
-        bytes4 interfaceId
-    )
+    function supportsInterface(bytes4 interfaceId)
         public
         view
         virtual
@@ -279,10 +279,12 @@ abstract contract MajorityVotingBase is
     }
 
     /// @inheritdoc IMajorityVoting
-    function getVoteOption(
-        uint256 _proposalId,
-        address _voter
-    ) public view virtual returns (VoteOption) {
+    function getVoteOption(uint256 _proposalId, address _voter)
+        public
+        view
+        virtual
+        returns (VoteOption)
+    {
         return proposals[_proposalId].voters[_voter];
     }
 
@@ -374,9 +376,8 @@ abstract contract MajorityVotingBase is
     /// @return parameters The parameters of the proposal vote.
     /// @return tally The current tally of the proposal vote.
     /// @return actions The actions to be executed in the associated DAO after the proposal has passed.
-    function getProposal(
-        uint256 _proposalId
-    )
+    /// @return allowFailureMap The bit map representations of which actions are allowed to revert so tx still succeeds.
+    function getProposal(uint256 _proposalId)
         public
         view
         virtual
@@ -385,7 +386,8 @@ abstract contract MajorityVotingBase is
             bool executed,
             ProposalParameters memory parameters,
             Tally memory tally,
-            IDAO.Action[] memory actions
+            IDAO.Action[] memory actions,
+            uint256 allowFailureMap
         )
     {
         Proposal storage proposal_ = proposals[_proposalId];
@@ -395,6 +397,7 @@ abstract contract MajorityVotingBase is
         parameters = proposal_.parameters;
         tally = proposal_.tally;
         actions = proposal_.actions;
+        allowFailureMap = proposal_.allowFailureMap;
     }
 
     /// @notice Returns the current voting settings.
@@ -405,15 +408,18 @@ abstract contract MajorityVotingBase is
 
     /// @notice Updates the voting settings.
     /// @param _votingSettings The new voting settings.
-    function updateVotingSettings(
-        VotingSettings calldata _votingSettings
-    ) external virtual auth(UPDATE_VOTING_SETTINGS_PERMISSION_ID) {
+    function updateVotingSettings(VotingSettings calldata _votingSettings)
+        external
+        virtual
+        auth(UPDATE_VOTING_SETTINGS_PERMISSION_ID)
+    {
         _updateVotingSettings(_votingSettings);
     }
 
     /// @notice Creates a new majority voting proposal.
     /// @param _metadata The metadata of the proposal.
     /// @param _actions The actions that will be executed after the proposal passes.
+    /// @param _allowFailureMap Allows proposal to succeed even if an action reverts. Uses bitmap representation. If the bit at index `x` is 1, the tx succeeds even if the action at `x` failed. Passing 0 will be treated as atomic execution.
     /// @param _startDate The start date of the proposal vote. If 0, the current timestamp is used and the vote starts immediately.
     /// @param _endDate The end date of the proposal vote. If 0, `_startDate + minDuration` is used.
     /// @param _voteOption The chosen vote option to be casted on proposal creation.
@@ -422,6 +428,7 @@ abstract contract MajorityVotingBase is
     function createProposal(
         bytes calldata _metadata,
         IDAO.Action[] calldata _actions,
+        uint256 _allowFailureMap,
         uint64 _startDate,
         uint64 _endDate,
         VoteOption _voteOption,
@@ -444,7 +451,12 @@ abstract contract MajorityVotingBase is
     function _execute(uint256 _proposalId) internal virtual {
         proposals[_proposalId].executed = true;
 
-        _executeProposal(dao, _proposalId, proposals[_proposalId].actions);
+        _executeProposal(
+            dao,
+            _proposalId,
+            proposals[_proposalId].actions,
+            proposals[_proposalId].allowFailureMap
+        );
     }
 
     /// @notice Internal function to check if a voter can vote. It assumes the queried proposal exists.
@@ -543,10 +555,12 @@ abstract contract MajorityVotingBase is
     /// @param _end The end date of the proposal vote. If 0, `_start + minDuration` is used.
     /// @return startDate The validated start date of the proposal vote.
     /// @return endDate The validated end date of the proposal vote.
-    function _validateProposalDates(
-        uint64 _start,
-        uint64 _end
-    ) internal view virtual returns (uint64 startDate, uint64 endDate) {
+    function _validateProposalDates(uint64 _start, uint64 _end)
+        internal
+        view
+        virtual
+        returns (uint64 startDate, uint64 endDate)
+    {
         uint64 currentTimestamp = block.timestamp.toUint64();
 
         if (_start == 0) {