+++ packages/contracts/deploy/30_plugins/10_plugin-repos/00_create-plugins-repo.ts
@@ -15,7 +15,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     hre,
     'AddresslistVoting',
     'AddresslistVotingSetup',
-    [1, 0, 0],
     '0x'
   );
 
@@ -24,15 +23,14 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     hre,
     'TokenVoting',
     'TokenVotingSetup',
-    [1, 0, 0],
     '0x'
   );
 
   // AdminSetup
-  await createPluginRepo(hre, 'Admin', 'AdminSetup', [1, 0, 0], '0x');
+  await createPluginRepo(hre, 'Admin', 'AdminSetup', '0x');
 
   // MultisigSetup
-  await createPluginRepo(hre, 'Multisig', 'MultisigSetup', [1, 0, 0], '0x');
+  await createPluginRepo(hre, 'Multisig', 'MultisigSetup', '0x');
 };
 export default func;
 func.tags = ['Create_Register_Plugins'];