+++ packages/contracts/test/voting/addresslist-voting.ts
@@ -2,7 +2,7 @@ import {expect} from 'chai';
 import {ethers} from 'hardhat';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 
-import {DAO} from '../../typechain';
+import {AddresslistVoting, DAO} from '../../typechain';
 import {
   findEvent,
   DAO_EVENTS,
@@ -29,7 +29,7 @@ import {OZ_ERRORS} from '../test-utils/error';
 
 describe('AddresslistVoting', function () {
   let signers: SignerWithAddress[];
-  let voting: any;
+  let voting: AddresslistVoting;
   let dao: DAO;
   let dummyActions: any;
   let dummyMetadata: string;
@@ -82,7 +82,7 @@ describe('AddresslistVoting', function () {
       addresslistVotingFactoryBytecode,
       signers[0]
     );
-    voting = await AddresslistVotingFactory.deploy();
+    voting = (await AddresslistVotingFactory.deploy()) as AddresslistVoting;
 
     startDate = (await getTime()) + startOffset;
     endDate = startDate + votingSettings.minDuration;
@@ -170,7 +170,7 @@ describe('AddresslistVoting', function () {
       await expect(
         voting
           .connect(signers[1])
-          .createProposal(dummyMetadata, [], 0, 0, VoteOption.None, false)
+          .createProposal(dummyMetadata, [], 0, 0, 0, VoteOption.None, false)
       )
         .to.be.revertedWithCustomError(voting, 'ProposalCreationForbidden')
         .withArgs(signers[1].address);
@@ -178,7 +178,7 @@ describe('AddresslistVoting', function () {
       await expect(
         voting
           .connect(signers[0])
-          .createProposal(dummyMetadata, [], 0, 0, VoteOption.None, false)
+          .createProposal(dummyMetadata, [], 0, 0, 0, VoteOption.None, false)
       ).to.not.be.reverted;
     });
 
@@ -194,7 +194,7 @@ describe('AddresslistVoting', function () {
       await expect(
         voting
           .connect(signers[1])
-          .createProposal(dummyMetadata, [], 0, 0, VoteOption.None, false)
+          .createProposal(dummyMetadata, [], 0, 0, 0, VoteOption.None, false)
       )
         .to.be.revertedWithCustomError(voting, 'ProposalCreationForbidden')
         .withArgs(signers[1].address);
@@ -202,7 +202,7 @@ describe('AddresslistVoting', function () {
       await expect(
         voting
           .connect(signers[0])
-          .createProposal(dummyMetadata, [], 0, 0, VoteOption.None, false)
+          .createProposal(dummyMetadata, [], 0, 0, 0, VoteOption.None, false)
       ).to.not.be.reverted;
     });
 
@@ -219,6 +219,7 @@ describe('AddresslistVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           startDateInThePast,
           endDate,
           VoteOption.None,
@@ -245,6 +246,7 @@ describe('AddresslistVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           tooLateStartDate,
           endDate,
           VoteOption.None,
@@ -266,6 +268,7 @@ describe('AddresslistVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           startDate,
           tooEarlyEndDate,
           VoteOption.None,
@@ -290,6 +293,7 @@ describe('AddresslistVoting', function () {
           await voting.createProposal(
             dummyMetadata,
             dummyActions,
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -315,6 +319,7 @@ describe('AddresslistVoting', function () {
           await voting.createProposal(
             dummyMetadata,
             dummyActions,
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -333,9 +338,12 @@ describe('AddresslistVoting', function () {
         signers.slice(0, 10).map(s => s.address)
       );
 
+      const allowFailureMap = 1;
+
       let tx = await voting.createProposal(
         dummyMetadata,
         dummyActions,
+        allowFailureMap,
         0,
         0,
         VoteOption.None,
@@ -354,12 +362,14 @@ describe('AddresslistVoting', function () {
       expect(event.args.actions[0].to).to.equal(dummyActions[0].to);
       expect(event.args.actions[0].value).to.equal(dummyActions[0].value);
       expect(event.args.actions[0].data).to.equal(dummyActions[0].data);
+      expect(event.args.allowFailureMap).to.equal(allowFailureMap);
 
       const block = await ethers.provider.getBlock('latest');
 
       const proposal = await voting.getProposal(id);
       expect(proposal.open).to.be.true;
       expect(proposal.executed).to.be.false;
+      expect(proposal.allowFailureMap).to.equal(allowFailureMap);
       expect(proposal.parameters.snapshotBlock).to.equal(block.number - 1);
       expect(proposal.parameters.supportThreshold).to.equal(
         votingSettings.supportThreshold
@@ -401,6 +411,7 @@ describe('AddresslistVoting', function () {
         dummyActions,
         0,
         0,
+        0,
         VoteOption.Yes,
         false
       );
@@ -418,11 +429,13 @@ describe('AddresslistVoting', function () {
       expect(event.args.actions[0].to).to.equal(dummyActions[0].to);
       expect(event.args.actions[0].value).to.equal(dummyActions[0].value);
       expect(event.args.actions[0].data).to.equal(dummyActions[0].data);
+      expect(event.args.allowFailureMap).to.equal(0);
 
       const block = await ethers.provider.getBlock('latest');
       const proposal = await voting.getProposal(id);
       expect(proposal.open).to.be.true;
       expect(proposal.executed).to.be.false;
+      expect(proposal.allowFailureMap).to.equal(0);
       expect(proposal.parameters.snapshotBlock).to.equal(block.number - 1);
       expect(proposal.parameters.supportThreshold).to.equal(
         votingSettings.supportThreshold
@@ -449,6 +462,7 @@ describe('AddresslistVoting', function () {
         voting.createProposal(
           dummyMetadata,
           dummyActions,
+          0,
           startDate,
           endDate,
           VoteOption.Yes,
@@ -464,6 +478,7 @@ describe('AddresslistVoting', function () {
           await voting.createProposal(
             dummyMetadata,
             dummyActions,
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -490,6 +505,7 @@ describe('AddresslistVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -620,6 +636,7 @@ describe('AddresslistVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -814,6 +831,7 @@ describe('AddresslistVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -949,6 +967,7 @@ describe('AddresslistVoting', function () {
           dummyActions,
           0,
           0,
+          0,
           VoteOption.None,
           false
         );
@@ -1055,6 +1074,7 @@ describe('AddresslistVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -1194,6 +1214,7 @@ describe('AddresslistVoting', function () {
           dummyActions,
           0,
           0,
+          0,
           VoteOption.None,
           false
         );
@@ -1248,6 +1269,7 @@ describe('AddresslistVoting', function () {
         await voting.createProposal(
           dummyMetadata,
           dummyActions,
+          0,
           0,
           0,
           VoteOption.None,