+++ packages/contracts/test/voting/token-voting.ts
@@ -7,6 +7,7 @@ import {
   DAO,
   GovernanceERC20Mock,
   GovernanceERC20Mock__factory,
+  TokenVoting,
 } from '../../typechain';
 import {
   findEvent,
@@ -35,7 +36,7 @@ import {OZ_ERRORS} from '../test-utils/error';
 
 describe('TokenVoting', function () {
   let signers: SignerWithAddress[];
-  let voting: any;
+  let voting: TokenVoting;
   let dao: DAO;
   let governanceErc20Mock: GovernanceERC20Mock;
   let GovernanceERC20Mock: GovernanceERC20Mock__factory;
@@ -104,7 +105,7 @@ describe('TokenVoting', function () {
       tokenVotingFactoryBytecode,
       signers[0]
     );
-    voting = await TokenVotingFactory.deploy();
+    voting = (await TokenVotingFactory.deploy()) as TokenVoting;
 
     startDate = (await getTime()) + startOffset;
     endDate = startDate + votingSettings.minDuration;
@@ -223,6 +224,7 @@ describe('TokenVoting', function () {
           .createProposal(
             dummyMetadata,
             [],
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -236,6 +238,7 @@ describe('TokenVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           startDate,
           endDate,
           VoteOption.None,
@@ -266,6 +269,7 @@ describe('TokenVoting', function () {
           .createProposal(
             dummyMetadata,
             [],
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -281,6 +285,7 @@ describe('TokenVoting', function () {
           .createProposal(
             dummyMetadata,
             [],
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -307,7 +312,15 @@ describe('TokenVoting', function () {
       );
 
       await expect(
-        voting.createProposal(dummyMetadata, [], 0, 0, VoteOption.None, false)
+        voting.createProposal(
+          dummyMetadata,
+          [],
+          0,
+          0,
+          0,
+          VoteOption.None,
+          false
+        )
       ).to.be.revertedWithCustomError(voting, 'NoVotingPower');
     });
 
@@ -326,6 +339,7 @@ describe('TokenVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           startDateInThePast,
           endDate,
           VoteOption.None,
@@ -354,6 +368,7 @@ describe('TokenVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           tooLateStartDate,
           endDate,
           VoteOption.None,
@@ -377,6 +392,7 @@ describe('TokenVoting', function () {
         voting.createProposal(
           dummyMetadata,
           [],
+          0,
           startDate,
           tooEarlyEndDate,
           VoteOption.None,
@@ -403,6 +419,7 @@ describe('TokenVoting', function () {
           await voting.createProposal(
             dummyMetadata,
             dummyActions,
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -430,6 +447,7 @@ describe('TokenVoting', function () {
           await voting.createProposal(
             dummyMetadata,
             dummyActions,
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -448,11 +466,14 @@ describe('TokenVoting', function () {
         governanceErc20Mock.address
       );
 
+      const allowFailureMap = 1;
+
       await setBalances([{receiver: signers[0].address, amount: 10}]);
 
       let tx = await voting.createProposal(
         dummyMetadata,
         dummyActions,
+        allowFailureMap,
         0,
         0,
         VoteOption.None,
@@ -471,6 +492,7 @@ describe('TokenVoting', function () {
       expect(event.args.actions[0].to).to.equal(dummyActions[0].to);
       expect(event.args.actions[0].value).to.equal(dummyActions[0].value);
       expect(event.args.actions[0].data).to.equal(dummyActions[0].data);
+      expect(event.args.allowFailureMap).to.equal(allowFailureMap);
 
       const block = await ethers.provider.getBlock('latest');
 
@@ -478,6 +500,7 @@ describe('TokenVoting', function () {
 
       expect(proposal.open).to.equal(true);
       expect(proposal.executed).to.equal(false);
+      expect(proposal.allowFailureMap).to.equal(allowFailureMap);
       expect(proposal.parameters.supportThreshold).to.equal(
         votingSettings.supportThreshold
       );
@@ -520,6 +543,7 @@ describe('TokenVoting', function () {
         dummyActions,
         0,
         0,
+        0,
         VoteOption.Yes,
         false
       );
@@ -537,12 +561,14 @@ describe('TokenVoting', function () {
       expect(event.args.actions[0].to).to.equal(dummyActions[0].to);
       expect(event.args.actions[0].value).to.equal(dummyActions[0].value);
       expect(event.args.actions[0].data).to.equal(dummyActions[0].data);
+      expect(event.args.allowFailureMap).to.equal(0);
 
       const block = await ethers.provider.getBlock('latest');
 
       const proposal = await voting.getProposal(id);
       expect(proposal.open).to.equal(true);
       expect(proposal.executed).to.equal(false);
+      expect(proposal.allowFailureMap).to.equal(0);
       expect(proposal.parameters.supportThreshold).to.equal(
         votingSettings.supportThreshold
       );
@@ -573,6 +599,7 @@ describe('TokenVoting', function () {
         voting.createProposal(
           dummyMetadata,
           dummyActions,
+          0,
           startDate,
           endDate,
           VoteOption.Yes,
@@ -588,6 +615,7 @@ describe('TokenVoting', function () {
           await voting.createProposal(
             dummyMetadata,
             dummyActions,
+            0,
             startDate,
             endDate,
             VoteOption.None,
@@ -629,6 +657,7 @@ describe('TokenVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -755,6 +784,7 @@ describe('TokenVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -977,6 +1007,7 @@ describe('TokenVoting', function () {
             await voting.createProposal(
               dummyMetadata,
               dummyActions,
+              0,
               startDate,
               endDate,
               VoteOption.None,
@@ -1126,6 +1157,7 @@ describe('TokenVoting', function () {
           dummyActions,
           0,
           0,
+          0,
           VoteOption.None,
           false
         );
@@ -1231,6 +1263,7 @@ describe('TokenVoting', function () {
           dummyActions,
           0,
           0,
+          0,
           VoteOption.None,
           false
         );
@@ -1303,6 +1336,7 @@ describe('TokenVoting', function () {
           dummyActions,
           0,
           0,
+          0,
           VoteOption.None,
           false
         );
@@ -1365,6 +1399,7 @@ describe('TokenVoting', function () {
         await voting.createProposal(
           dummyMetadata,
           dummyActions,
+          0,
           0,
           0,
           VoteOption.None,
@@ -1452,6 +1487,7 @@ describe('TokenVoting', function () {
         await voting.createProposal(
           dummyMetadata,
           dummyActions,
+          0,
           0,
           0,
           VoteOption.None,